+++ designspace/src/lib.rs
@@ -50,10 +50,10 @@ impl Designspace {
             .expect("Couldn't open name table");
 
         for axis in self.axes.axis.iter() {
-            axes.push(axis.to_variation_axis_record(ix as u16)?);
+            axes.push(axis.to_variation_axis_record(ix)?);
 
             name.records
-                .push(NameRecord::windows_unicode(ix as u16, axis.name.clone()));
+                .push(NameRecord::windows_unicode(ix, axis.name.clone()));
 
             ix += 1;
             if axis.map.is_some() {
